+++ app/src/main/java/com/daniil/halushka/todoapp/presentation/screens/home/HomeScreenViewModel.kt
@@ -3,9 +3,9 @@ package com.daniil.halushka.todoapp.presentation.screens.home
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.daniil.halushka.todoapp.data.models.TodoItem
-import com.daniil.halushka.todoapp.domain.usecase.CountFinishedTodo
-import com.daniil.halushka.todoapp.domain.usecase.FinishTodo
-import com.daniil.halushka.todoapp.domain.usecase.ReceiveTodoList
+import com.daniil.halushka.todoapp.domain.usecases.CountFinishedTodo
+import com.daniil.halushka.todoapp.domain.usecases.FinishTodo
+import com.daniil.halushka.todoapp.domain.usecases.ReceiveTodoList
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
